refactor: threejs instance is run as a module
