+++ src/App.vue
@@ -1,157 +1,27 @@
 <template>
-  <div id="container"></div>
   <navigation-bar></navigation-bar>
   <section-landing></section-landing>
-  <!-- <section-projects></section-projects> -->
+  <section-projects></section-projects>
   <!-- <section-why></section-why> -->
   <!-- <section-contact></section-contact> -->
+  <div id="container"></div>
 </template>
 
 <script>
 import NavigationBar from "./components/NavigationBar.vue";
 import SectionLanding from "./sections/Landing.vue";
-import SectionWhy from "./sections/Why.vue";
 import SectionProjects from "./sections/Projects.vue";
-import SectionContact from "./sections/Contact.vue";
-import * as THREE from "three";
-import Stats from "three/examples/jsm/libs/stats.module.js"
+import { useThree } from "./modules/useThreeBackground";
 
 export default {
   name: "App",
   components: {
     NavigationBar,
     SectionLanding,
-    SectionWhy,
     SectionProjects,
-    SectionContact,
   },
   mounted() {
-    let camera, scene, renderer, stats, material;
-    let mouseX = 0,
-      mouseY = 0;
-
-    let windowHalfX = window.innerWidth / 2;
-    let windowHalfY = window.innerHeight / 2;
-
-    init();
-    animate();
-
-    function init() {
-      camera = new THREE.PerspectiveCamera(
-        55,
-        window.innerWidth / window.innerHeight,
-        2,
-        2000
-      );
-      camera.position.z = 1000;
-
-      scene = new THREE.Scene();
-      scene.fog = new THREE.FogExp2(0x000000, 0.001);
-
-      const geometry = new THREE.BufferGeometry();
-      const vertices = [];
-
-      const sprite = new THREE.TextureLoader().load(
-        "light.png"
-      );
-
-      for (let i = 0; i < 10000; i++) {
-        const x = 3000 * Math.random() - 1000;
-        const y = 3000 * Math.random() - 1000;
-        const z = 3000 * Math.random() - 1000;
-
-        vertices.push(x, y, z);
-      }
-
-      geometry.setAttribute(
-        "position",
-        new THREE.Float32BufferAttribute(vertices, 3)
-      );
-
-      material = new THREE.PointsMaterial({
-        size: 40,
-        sizeAttenuation: true,
-        map: sprite,
-        alphaTest: 0.5,
-        transparent: true,
-      });
-      material.color.setRGB(0.2, 0, 1);
-      // material.color.setRGB(1, 1, 1);
-
-      const particles = new THREE.Points(geometry, material);
-      scene.add(particles);
-
-      //
-
-      renderer = new THREE.WebGLRenderer();
-      renderer.setPixelRatio(window.devicePixelRatio);
-      renderer.setSize(window.innerWidth, window.innerHeight);
-      document.querySelector("#container").appendChild(renderer.domElement);
-
-      //
-
-      /* stats = new Stats();
-      document.body.appendChild(stats.dom); */
-
-      //
-
-/*       const gui = new GUI();
-
-      gui.add(material, "sizeAttenuation").onChange(function () {
-        material.needsUpdate = true;
-      });
-
-      gui.open(); */
-
-      //
-
-      document.body.style.touchAction = "none";
-      document.body.addEventListener("pointermove", onPointerMove);
-
-      //
-
-      window.addEventListener("resize", onWindowResize);
-    }
-
-    function onWindowResize() {
-      windowHalfX = window.innerWidth / 2;
-      windowHalfY = window.innerHeight / 2;
-
-      camera.aspect = window.innerWidth / window.innerHeight;
-      camera.updateProjectionMatrix();
-
-      renderer.setSize(window.innerWidth, window.innerHeight);
-    }
-
-    function onPointerMove(event) {
-      if (event.isPrimary === false) return;
-
-      mouseX = event.clientX - windowHalfX;
-      mouseY = event.clientY - windowHalfY;
-    }
-
-    //
-
-    function animate() {
-      requestAnimationFrame(animate);
-
-      render();
-      // stats.update();
-    }
-
-    function render() {
-      const time = Date.now() * 0.00005;
-
-      camera.position.x += (mouseX - camera.position.x) * 0.05;
-      camera.position.y += (-mouseY - camera.position.y) * 0.05;
-
-      camera.lookAt(scene.position);
-
-      // const h = ((360 * (1.0 + time)) % 360) / 360;
-      // material.color.setHSL(h, 0.5, 0.5);
-
-      renderer.render(scene, camera);
-    }
+    useThree("#container");
   },
 };
 </script>
@@ -164,6 +34,6 @@ export default {
   -moz-osx-font-smoothing: grayscale;
 }
 #container {
-  @apply h-[200px] w-[200px] fixed z-[1];
+  @apply fixed z-0 top-0 left-0 right-0 bottom-0;
 }
 </style>
